set explicit versions for logging components
Elasticsearch has made major changes in version 8.x which is used by latest Helm package they maintain. These major changes are related to authentication and by default do not work with our Helm value files. So setting version to latest version of 7th release which is working as expected as of creation of Playground script.