Use explicit, working versions for logging stack
Elasticsearch has made major changes in version 8.x, which is automatically pulled by Helm we run inside the setup.sh script. These major changes eventually lead to failing deployment of ES, Kibana and Fluentd connector Pods. Our Helm values for logging components should be adjusted to make everything to work again.

Quick fix is proposed in this PR to still use that suitable for our version which is was probably the latest version back then when this script was developed initially.